Split formatters into _internal/formatting
